Feat: implements delete user and improves response messages

+++ src/modules/users/users.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, Param, Patch, Post, Query, UseGuards, Put } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Param, Patch, Post, Query, UseGuards, Put, Req, InternalServerErrorException, BadRequestException } from '@nestjs/common';
 import { Role } from '@prisma/client';
 import { HasRole } from '@src/common/decorators/has-role.decorator';
 import { PageQuery } from '@src/common/http/page-query.type';
@@ -11,16 +11,16 @@ import { RolesGuard } from '../auth/guards/roles.guard';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UsersService } from './users.service';
+import { AuthPayload } from '../auth/dtos/auth-payload.dto';
 
-@Controller('users')
 
 // @UseGuards(RolesGuard)
 // @UseGuards(JwtAuthGuard)
+@Controller('users')
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
 
   @Get(':id')
-  @HasRole()
   async findOne(@Param('id', IsId) id: number): Promise<User> {
     return await this.usersService.findOne(+id);
   }
@@ -32,6 +32,7 @@ export class UsersController {
 
   @Post()
   @HasRole(Role.ADMIN)
+  @UseGuards(RolesGuard)
   @UseGuards(JwtAuthGuard)
   async create(@Body() createUserDto: CreateUserDto): Promise<User> {
     return await this.usersService.create(createUserDto);
@@ -39,6 +40,7 @@ export class UsersController {
 
   @Patch(':id')
   @HasRole(Role.ADMIN)
+  @UseGuards(RolesGuard)
   @UseGuards(JwtAuthGuard)
   async update(@Param('id', IsId) id: number, @Body() updateUserDto: UpdateUserDto): Promise<User> {
     return await this.usersService.update(id, updateUserDto);
@@ -46,12 +48,19 @@ export class UsersController {
 
   @Delete(':id')
   @HasRole(Role.ADMIN)
+  @UseGuards(RolesGuard)
   @UseGuards(JwtAuthGuard)
-  remove(@Param('id', IsId) id: number) {
+  remove(@Param('id', IsId) id: number, @Req() request: Request) {
+    const authPayload: AuthPayload = (request as any).user
+
+    if (!authPayload.id) throw new InternalServerErrorException('Auth payload Vazio') 
+
+    if (Number(authPayload.id) == id) throw new BadRequestException('Usuário não pode ser removido')
+
     return this.usersService.remove(id);
   }
 
-  @Get()
+  @Get('/pages/page')
   async findAllPaginated(
     @Query(PageQueryPipe) query: PageQuery, 
   ): Promise<Page<User>> {
